modify based on feedbacks

+++ src/main/java/com/zoomers/GameSetMatch/repository/TournamentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TournamentRepository extends JpaRepository<Tournament, Long> {
+public interface TournamentRepository extends JpaRepository<Tournament, Integer> {
     List<Tournament> findByStatus(int status);
 }
